feat: access token has a dedicated impl
